add test coverage for P2445R1 (`std::forward_like`)
